+++ CHANGELOG.md
@@ -3,6 +3,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
+# [0.3.0] - 2016-11-21
+### Added
+- fourth release
+Added the possibility to tweak your queries by adding --period in your parameters.
+By default --period will be 10 minutes to work with previous versions
+
 # [0.2.1] - 2016-11-21
 ### Added
 - third release
@@ -21,4 +27,5 @@ Leaving 5 minutes to the data to consolidate.
 
 [0.1.0]: https://github.com/pliyosenpai/sensu-plugins-influxdb-metrics-checker/0.1.0...0.2.0
 [0.2.0]: https://github.com/pliyosenpai/sensu-plugins-influxdb-metrics-checker/0.2.0...0.2.1
-[0.2.1]: https://github.com/pliyosenpai/sensu-plugins-influxdb-metrics-checker/0.2.1...HEAD
+[0.2.1]: https://github.com/pliyosenpai/sensu-plugins-influxdb-metrics-checker/0.2.1...0.3.0
+[0.3.0]: https://github.com/pliyosenpai/sensu-plugins-influxdb-metrics-checker/0.3.0...HEAD